The sediment model has been written and added to the system.
Sediment is now a valid flow-model name from commandline.

When flow-model="Sediment", aries-span-bridge now fetches the
Hydrosheds layer, downsamples it to match the other layers passed in,
and adds the final matrix to flow-layers.

core uses aggregate-flow-dirs to downsample Hydrosheds and rv-average
for all other layers.

Improved downsample-matrix to use interleaved sampling if the original
rows and cols aren't evenly divisible by the downscaling-factor.
Downscaling-factor may now be any number.  This is allowed by
commandline and core's parameter tests and is documented in the
toplevel README.

The carbon and sediment models both store simply the originating
source or sink ids in their carrier routes rather than a bitpacked
path between the source/sink and use locations.

worldgen/make-random-layer-map now accepts 0-arity closures in its
name-to-type-map as well as the values :discrete and :continuous.