Optimization of clj-span.carbon-model.
The carbon model was carefully reorganized to make full use of
laziness (for memory savings).  However, the storage space for a
single world still scales very poorly as S*K*U (the number of non-zero
Source, sinK, and Use points respectively).  I need a better way of
storing this information, but this will require substantial
modification of clj-span.analyzer.